+++ components/Layout/LayoutStyle.tsx
@@ -1,0 +1,7 @@
+import styled from 'styled-components';
+
+export const Wrapper = styled.div``;
+
+export const Main = styled.div`
+  margin-top: 3.75rem;
+`;
