+++ pages/index.tsx
@@ -1,4 +1,4 @@
-import type { NextPage, NextApiRequest, NextApiResponse } from 'next';
+import type { NextPage } from 'next';
 import { useState } from 'react';
 import { getCookie } from 'cookies-next';
 import useSWR from 'swr';
@@ -7,12 +7,12 @@ import { getUserData } from '../lib/user';
 import { getContentsList } from '../lib/content';
 import { URI } from '../config';
 
-import List from '../components/List';
-import Banner from '../components/Banner';
-import Layout from '../components/Layout';
+import List from '../components/List/List';
+import Banner from '../components/Banner/Banner';
+import Layout from '../components/Layout/Layout';
 import PageBtns from '../components/PageBtns/PageBtns';
 
-const Home: NextPage = ({ userData }: any) => {
+const Main: NextPage = ({ userData }: any) => {
   const [currentStart, setCurrentStart] = useState(1);
   const { data, error } = useSWR(
     `${URI}/v1/contents?isActive=true&_start=${currentStart}&_limit=5&_sort=createdAt `,
@@ -38,4 +38,4 @@ export const getServerSideProps = async (props: any) => {
   return { props: { userData: userData.data } };
 };
 
-export default Home;
+export default Main;
